Add files via upload
With the impact of climate change in India, majority of the agricultural crops are being badly 
affected in terms of their performance over a period of last two decades. Predicting the crop 
yield well ahead of its harvest would help the farmers for taking appropriate measures for
marketing and storage. Several methods of predicting and modeling crop yields have been 
developed in the past with varying rate of success, as these don’t take into account 
characteristics of the weather, and are mostly empirical. In the present study, a predictive 
model has been developed for predicting the crop yield (production).Multiple Linear 
Regression algorithm is used to find out the crop yields of selected crops. The features which 
we are using to predict the yield are area, soil ph, temperature, rainfall, humidity, nitrogen, and 
potassium and phosphorus contents in soil