Add files via upload
FTP powers one of the fundamental Internet functions and is the prescribed method for the 
transfer of files between computers. It is also the easiest and most secure way to exchange files 
over the Internet. The File Transfer Protocol is one of the oldest and widely used protocols on the 
Internet. The protocol's primary character set, bit ASCII, has served the protocol well through the 
early growth years of the Internet. However, as the Internet becomes more global, there is a need 
to support character sets beyond 7 bit ASCII. This document addresses the internationalization 
(I18n) of FTP, which includes supporting the multiple character sets and languages found 
throughout the Internet community. This is achieved by extending the FTP specification and 
giving recommendations for proper internationalization support.