Add files via upload
Computers, on the other hand, require effort and powerful computation and complex algorithms to attempt to recognize correctly patterns and regions where a possible object might be. Object detection and recognition are two main ways that have been implemented over multiple decades that are at the center of Computer Vision systems at the moment. These approaches are presented with challenges such as scale, occlusion, viewpoint, illumination or background clutter, all issues that have been attempted as research topics that provided functionality that led to the introduction of Neural Networks and Convolutional Neural Networks (CNN).
1.2 Aim and Objective
The main aim and objectives are as below
•	To get object like plane, car, etc. recognised.
•	To classify the image into a proper category.
•	Built a model which can perform both the function in a mannered way.