Add files via upload
Vehicle services management system is the project in which we are making the web- based 
database for the vehicle shops to store the information of the customers, staff and the services 
taken by each customer. In which we are creating the interface for the users to store the 
information in the efficient manner and it be easy for user to search it.
The information store in the database here mysql is used and connected using the web-based 
interface as front-end. We are making this database using the Sql language and interface using 
html, php, css and Javascript language.
So, it allows the users to add the details of customers, staffs, vehicle services and the amount 
related to service done by each customer and store all the information in the database for the 
record purpose. The vehicle services management system helps the user of vehicle services 
centre to efficiently maintain the record and generate the bill accordingly.